requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
